+++ source/isaaclab_ovphysx/isaaclab_ovphysx/assets/articulation/articulation.py
@@ -2181,7 +2181,7 @@ class Articulation(BaseArticulation):
         self._data._body_com_pose_b.timestamp = self._data._sim_timestamp
         self._data._reset_body_com_pose_b_dependents()
         cpu_env_ids = self._get_cpu_env_ids(env_ids)
-        wp.copy(self.data._cpu_body_coms, self._data._body_com_pose_b.data)
+        wp.copy(self.data._cpu_body_coms, self._data._body_com_pose_b.data.view(wp.float32))
         self._root_view.set_attribute(TT.BODY_COM_POSE, self.data._cpu_body_coms, indices=cpu_env_ids)
 
     def set_coms_mask(
@@ -2224,7 +2224,7 @@ class Articulation(BaseArticulation):
         )
         self._data._body_com_pose_b.timestamp = self._data._sim_timestamp
         self._data._reset_body_com_pose_b_dependents()
-        wp.copy(self.data._cpu_body_coms, self._data._body_com_pose_b.data)
+        wp.copy(self.data._cpu_body_coms, self._data._body_com_pose_b.data.view(wp.float32))
         self._root_view.set_attribute(
             TT.BODY_COM_POSE, self.data._cpu_body_coms, mask=self._get_cpu_env_mask(env_mask_wp)
         )